Create dist.py
Calculates distances and then writes them to a csv file. When many distances are needed, an input .txt file can be made containing 
atom #1, atom #2, y (to continue or n when at the last distance needed) all in a column. For example a text file could contain:
y
1
2
y
2
3
n
The output will give the distance between atom 1 and 2 and atom 3 and 4. The atom numbers are based off of the xyz file. The first atom listed in the xyz file is atom 1. 
To run the input file simple do something like "python3 ./dist.py xyzfile.xyz < input.txt "